+++ core/tests/test_views.py
@@ -1,5 +1,6 @@
 import http
 from unittest.mock import call, patch, PropertyMock
+from urllib.parse import urljoin
 
 import requests
 
@@ -16,7 +17,7 @@ from core import helpers, views
 from core.tests.helpers import create_response
 from casestudy import casestudies
 
-from directory_constants import slugs
+from directory_constants import slugs, urls
 
 
 def test_exopps_redirect(client):
@@ -1038,7 +1039,13 @@ def test_international_header_on(mock_get_page, client, settings):
 
 def test_international_trade_redirect_home(client):
     url = reverse('international-trade-home')
-    match = resolve(url)
-    import pdb; pdb.set_trace()
     response = client.get(url)
     assert response.status_code == 302
+    assert response.url == urljoin(urls.GREAT_INTERNATIONAL, 'trade/')
+
+
+def test_international_trade_redirect(client):
+    url = reverse('international-trade', kwargs={'path': '/foo/bar'})
+    response = client.get(url)
+    assert response.status_code == 302
+    assert response.url == urljoin(urls.GREAT_INTERNATIONAL, 'trade/incoming/foo/bar/')